// web/src/lib/types/store.ts
export interface Store {
  address: string
  currency: string
  currency_rate: string
  email: string
  fe_lang: string
  fe_locale: string
  fe_theme: string
  feature_floor: string
  feature_room: string
  feature_table: string
  logo: string
  name: string
  phone: string
  pos_type: string
  service_category: string
  service_rate: string
  tax_category: string
  tax_rate: string
}